Moved the session's page cache to it's own class
We need to improve cache configurability for handling Ajax/partial page
renders. This also lays the ground in general for more advanced tuning
of the lifecycle for certain pages. For example, you might want to
dispose a heavy page you're "done with" early, or in other cases, want
some instances to stay around for a while