Eliminated "languages" as an attribute of the template parser
Languages were initially passed to the template parser, only to allow it
to pass them on to the construction of dynamic elements (which in the
current environment, allow locale specific construction). Since dynamic
element construction will soon no longer be the responsibility of the
parser, we don't need this here any more.